Move AggregatedProblems to a more appropriate place
