+++ std-bits/table/src/main/java/org/enso/table/data/table/Table.java
@@ -16,7 +16,7 @@ import org.enso.table.data.table.join.IndexJoin;
 import org.enso.table.data.table.join.JoinCondition;
 import org.enso.table.data.table.join.JoinResult;
 import org.enso.table.data.table.join.JoinStrategy;
-import org.enso.table.data.table.problems.AggregatedProblems;
+import org.enso.table.problems.AggregatedProblems;
 import org.enso.table.error.UnexpectedColumnTypeException;
 import org.enso.table.operations.Distinct;
 import org.enso.table.util.NameDeduplicator;
