+++ std-bits/table/src/main/java/org/enso/table/data/table/join/JoinResult.java
@@ -1,6 +1,6 @@
 package org.enso.table.data.table.join;
 
-import org.enso.table.data.table.problems.AggregatedProblems;
+import org.enso.table.problems.AggregatedProblems;
 import org.graalvm.collections.Pair;
 
 import java.util.List;
